Move tests from HiveCommandSuite
